requirements.txt
Flask
pandas
nltk
scikit-learn
python-telegram-bot
requests
